+++ services/web/server/src/simcore_service_webserver/login/_controller/rest/change.py
@@ -10,16 +10,16 @@ from servicelib.mimetype_constants import MIMETYPE_APPLICATION_JSON
 from servicelib.request_keys import RQT_USERID_KEY
 from simcore_postgres_database.utils_users import UsersRepo
 
-from ..._meta import API_VTAG
-from ...db.plugin import get_database_engine
-from ...products import products_web
-from ...products.models import Product
-from ...security.api import check_password, encrypt_password
-from ...users import api as users_service
-from ...utils import HOUR
-from ...utils_rate_limiting import global_rate_limit_route
-from .. import _confirmation_service
-from .._constants import (
+from ...._meta import API_VTAG
+from ....db.plugin import get_database_engine
+from ....products import products_web
+from ....products.models import Product
+from ....security.api import check_password, encrypt_password
+from ....users import api as users_service
+from ....utils import HOUR
+from ....utils_rate_limiting import global_rate_limit_route
+from ... import _confirmation_service
+from ..._constants import (
     MSG_CANT_SEND_MAIL,
     MSG_CHANGE_EMAIL_REQUESTED,
     MSG_EMAIL_SENT,
@@ -27,17 +27,17 @@ from .._constants import (
     MSG_PASSWORD_CHANGED,
     MSG_WRONG_PASSWORD,
 )
-from .._emails_service import get_template_path, send_email_from_template
-from .._login_repository_legacy import AsyncpgStorage, get_plugin_storage
-from .._login_service import (
+from ..._emails_service import get_template_path, send_email_from_template
+from ..._login_repository_legacy import AsyncpgStorage, get_plugin_storage
+from ..._login_service import (
     ACTIVE,
     CHANGE_EMAIL,
     flash_response,
     validate_user_status,
 )
-from .._models import InputSchema, create_password_match_validator
-from ..decorators import login_required
-from ..settings import LoginOptions, get_plugin_options
+from ..._models import InputSchema, create_password_match_validator
+from ...decorators import login_required
+from ...settings import LoginOptions, get_plugin_options
 
 _logger = logging.getLogger(__name__)
